split off intense_traffic_test from distbench_test_sequencer_test
